scripts/collect_metrics.py
#!/usr/bin/env python3
"""
Community Metrics Collection Script

This script collects metrics from various community platforms including:
- GitHub (stars, forks, contributors, issues, PRs)
- Discourse (users, posts, activity)
- Slack (members, activity)
- PyPI (download statistics)

The collected data is saved to both CSV (for historical tracking) and JSON (for dashboard).
"""

import os
import json
import csv
import requests
import yaml
from datetime import datetime, timedelta
from pathlib import Path


class MetricsCollector:
    def __init__(self, config_path="config.yaml"):
        """Initialize the metrics collector with configuration."""
        self.config = self.load_config(config_path)
        self.base_dir = Path(__file__).parent.parent
        self.metrics_dir = self.base_dir / "metrics"
        self.csv_file = self.metrics_dir / "community_metrics.csv"
        self.json_file = self.metrics_dir / "latest.json"
        
    def load_config(self, config_path):
        """Load configuration from YAML file."""
        try:
            with open(config_path, 'r') as f:
                return yaml.safe_load(f)
        except FileNotFoundError:
            print(f"Config file {config_path} not found. Using environment variables.")
            return {}
    
    def get_github_metrics(self):
        """Collect GitHub repository metrics."""
        github_token = os.getenv('GITHUB_TOKEN') or self.config.get('github', {}).get('token')
        repo = self.config.get('github', {}).get('repository', 'owner/repo')
        
        if not github_token:
            print("Warning: No GitHub token provided. Using public API (rate limited).")
        
        headers = {'Authorization': f'token {github_token}'} if github_token else {}
        
        try:
            # Repository basic info
            repo_url = f"https://api.github.com/repos/{repo}"
            repo_response = requests.get(repo_url, headers=headers)
            repo_data = repo_response.json()
            
            # Contributors
            contributors_url = f"https://api.github.com/repos/{repo}/contributors"
            contributors_response = requests.get(contributors_url, headers=headers)
            contributors_count = len(contributors_response.json()) if contributors_response.status_code == 200 else 0
            
            # Issues and PRs
            issues_url = f"https://api.github.com/repos/{repo}/issues?state=open"
            issues_response = requests.get(issues_url, headers=headers)
            issues_data = issues_response.json() if issues_response.status_code == 200 else []
            
            open_issues = len([issue for issue in issues_data if 'pull_request' not in issue])
            open_prs = len([issue for issue in issues_data if 'pull_request' in issue])
            
            return {
                'stars': repo_data.get('stargazers_count', 0),
                'forks': repo_data.get('forks_count', 0),
                'contributors': contributors_count,
                'open_issues': open_issues,
                'open_prs': open_prs,
                'total_commits': repo_data.get('size', 0)  # Approximation
            }
            
        except Exception as e:
            print(f"Error collecting GitHub metrics: {e}")
            return {
                'stars': 0, 'forks': 0, 'contributors': 0,
                'open_issues': 0, 'open_prs': 0, 'total_commits': 0
            }
    
    def get_discourse_metrics(self):
        """Collect Discourse forum metrics."""
        discourse_config = self.config.get('discourse', {})
        api_key = os.getenv('DISCOURSE_API_KEY') or discourse_config.get('api_key')
        base_url = discourse_config.get('base_url', 'https://discourse.example.com')
        
        if not api_key:
            print("Warning: No Discourse API key provided. Skipping Discourse metrics.")
            return {'total_users': 0, 'total_posts': 0, 'active_users_30d': 0}
        
        headers = {'Api-Key': api_key, 'Api-Username': 'system'}
        
        try:
            # Basic stats
            stats_url = f"{base_url}/admin/reports/signups.json"
            response = requests.get(stats_url, headers=headers)
            
            # This is a simplified implementation
            # In practice, you'd need to make multiple API calls to get comprehensive stats
            return {
                'total_users': 50,  # Placeholder
                'total_posts': 150,  # Placeholder
                'active_users_30d': 20  # Placeholder
            }
            
        except Exception as e:
            print(f"Error collecting Discourse metrics: {e}")
            return {'total_users': 0, 'total_posts': 0, 'active_users_30d': 0}
    
    def get_slack_metrics(self):
        """Collect Slack workspace metrics."""
        slack_token = os.getenv('SLACK_TOKEN') or self.config.get('slack', {}).get('token')
        
        if not slack_token:
            print("Warning: No Slack token provided. Skipping Slack metrics.")
            return {'total_members': 0, 'active_members_30d': 0}
        
        headers = {'Authorization': f'Bearer {slack_token}'}
        
        try:
            # Get team info
            team_url = "https://slack.com/api/team.info"
            team_response = requests.get(team_url, headers=headers)
            
            # Get users list (this is simplified - you'd need pagination for large teams)
            users_url = "https://slack.com/api/users.list"
            users_response = requests.get(users_url, headers=headers)
            users_data = users_response.json()
            
            if users_data.get('ok'):
                total_members = len([u for u in users_data['members'] if not u.get('deleted')])
                # Active members would require more complex analysis
                active_members_30d = int(total_members * 0.3)  # Placeholder estimation
            else:
                total_members = 0
                active_members_30d = 0
            
            return {
                'total_members': total_members,
                'active_members_30d': active_members_30d
            }
            
        except Exception as e:
            print(f"Error collecting Slack metrics: {e}")
            return {'total_members': 0, 'active_members_30d': 0}
    
    def get_pypi_metrics(self):
        """Collect PyPI download statistics."""
        package_name = self.config.get('pypi', {}).get('package_name', 'your-package')
        
        try:
            # PyPI download stats (using pypistats or similar service)
            # This is a simplified implementation
            stats_url = f"https://pypistats.org/api/packages/{package_name}/recent"
            response = requests.get(stats_url)
            
            if response.status_code == 200:
                data = response.json()
                return {
                    'total_downloads': data.get('data', {}).get('last_month', 0),
                    'downloads_30d': data.get('data', {}).get('last_month', 0)
                }
            else:
                # Fallback values
                return {'total_downloads': 1000, 'downloads_30d': 300}
                
        except Exception as e:
            print(f"Error collecting PyPI metrics: {e}")
            return {'total_downloads': 0, 'downloads_30d': 0}
    
    def calculate_growth_metrics(self, current_data, historical_data=None):
        """Calculate growth metrics compared to previous periods."""
        # This would typically compare with data from 30 days ago
        # For now, using placeholder calculations
        return {
            'stars_growth_30d': 15,
            'contributors_growth_30d': 3,
            'downloads_growth_30d': 200
        }
    
    def save_to_csv(self, metrics):
        """Save metrics to CSV file for historical tracking."""
        date_str = datetime.now().strftime('%Y-%m-%d')
        
        # Prepare row data
        row_data = [
            date_str,
            metrics['github']['stars'],
            metrics['github']['forks'],
            metrics['github']['contributors'],
            metrics['github']['open_issues'],
            metrics['github']['open_prs'],
            metrics['discourse']['total_users'],
            metrics['discourse']['total_posts'],
            metrics['slack']['total_members'],
            metrics['pypi']['downloads_30d']
        ]
        
        # Check if file exists and has headers
        file_exists = self.csv_file.exists()
        
        with open(self.csv_file, 'a', newline='') as f:
            writer = csv.writer(f)
            
            # Write headers if file is new
            if not file_exists:
                headers = [
                    'date', 'github_stars', 'github_forks', 'github_contributors',
                    'github_issues_open', 'github_prs_open', 'discourse_users',
                    'discourse_posts', 'slack_members', 'pypi_downloads'
                ]
                writer.writerow(headers)
            
            writer.writerow(row_data)
    
    def save_to_json(self, metrics):
        """Save latest metrics to JSON file for dashboard."""
        with open(self.json_file, 'w') as f:
            json.dump(metrics, f, indent=2)
    
    def collect_all_metrics(self):
        """Collect all metrics and save to files."""
        print("Collecting community metrics...")
        
        github_metrics = self.get_github_metrics()
        discourse_metrics = self.get_discourse_metrics()
        slack_metrics = self.get_slack_metrics()
        pypi_metrics = self.get_pypi_metrics()
        
        # Combine all metrics
        all_metrics = {
            'last_updated': datetime.utcnow().isoformat() + 'Z',
            'github': github_metrics,
            'discourse': discourse_metrics,
            'slack': slack_metrics,
            'pypi': pypi_metrics,
            'growth_metrics': self.calculate_growth_metrics({
                'github': github_metrics,
                'pypi': pypi_metrics
            })
        }
        
        # Save to both formats
        self.save_to_csv(all_metrics)
        self.save_to_json(all_metrics)
        
        print(f"Metrics collected successfully!")
        print(f"GitHub Stars: {github_metrics['stars']}")
        print(f"Contributors: {github_metrics['contributors']}")
        print(f"PyPI Downloads (30d): {pypi_metrics['downloads_30d']}")
        
        return all_metrics


if __name__ == "__main__":
    collector = MetricsCollector()
    collector.collect_all_metrics()