requirements.txt
# Python dependencies for community metrics collection and dashboard generation
requests>=2.28.0
PyYAML>=6.0
pandas>=1.5.0
jinja2>=3.1.0

# Optional dependencies for advanced features
matplotlib>=3.6.0  # For generating static charts (if needed)
plotly>=5.11.0     # For interactive visualizations (if needed)
beautifulsoup4>=4.11.0  # For web scraping (if needed)

# Development dependencies
flake8>=5.0.0      # Code linting
black>=22.0.0      # Code formatting
pytest>=7.0.0      # Testing framework